Refactor codes and extract some components form feed component and fix loading state get posts and get post by user id
