+++ src/components/home-posts/HomePosts.scss
@@ -1,0 +1,4 @@
+.home-posts {
+  border: solid 1px red;
+  background-color: pink;
+}
