+++ src/components/profile-posts/ProfilePosts.scss
@@ -1,0 +1,4 @@
+.profile-posts {
+  border: solid 1px black;
+  background-color: aqua;
+}
